build: Fix BOARD_CUSTOM_BOOTIMG_MK for recovery target
Change-Id: I3daf0b6288421c3880c8b3a71c15bafb87ae3df3

-------------------------------------------------------------------------------------------
build: Actually build the recovery.img if !BOARD_CUSTOM_BOOTIMG_MK

Fix the BOARD_CUSTOM_BOOTIMG_MK if and call build of recovery.img
also removed unused variable RECOVERYIMAGE_EXTRA_DEPS until associated
commit is merged

Change-Id: I94f5f26bc01e03d761ccc260342f330a4a06b141

-------------------------------------------------------------------------------------------
build: Fix check on BOARD_CUSTOM_BOOTIMG_MK

In the current code, '!' is added as prefix to the string contained in
$(BOARD_CUSTOM_BOOTIMG_MK). This means that !$(BOARD_CUSTOM_BOOTIMG_MK)
is never an empty string and therefore the condition is always true.

The if function has a third optional argument that is evalueted when the
condition is false, so use it to fix the current code.

Change-Id: If1e459f6c40f655b47a12b77761921c943964e3b
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>